requirements.txt
streamlit==1.28.1
numpy==1.24.3
opencv-python==4.8.0.74
Pillow==10.0.0
scikit-learn==1.3.0
matplotlib==3.7.2
scipy==1.11.2
pandas==2.0.3
streamlit-drawable-canvas==0.2.2
pytesseract==0.3.10
